Introduce tweet filter abstraction
