+++ TwitterEtl.Worker/ITweetFilter.cs
@@ -1,0 +1,7 @@
+using System.Threading.Tasks;
+using Tweetinvi.Models;
+
+public interface ITweetFilter
+{
+    Task<bool> ShouldProcessAsync(ITweet tweet);
+}
