+++ TwitterEtl.Worker/Program.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Hosting;
 IHost host = Host.CreateDefaultBuilder(args)
     .ConfigureServices(services =>
     {
+        services.AddSingleton<ITweetFilter, DefaultTweetFilter>();
         services.AddHostedService<Worker>();
     })
     .Build();
