+++ TwitterEtl.Worker/Worker.cs
@@ -3,31 +3,24 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Tweetinvi;
 using Tweetinvi.Models;
-using Azure.AI.OpenAI;
 
 public class Worker : BackgroundService
 {
     private readonly ILogger<Worker> _logger;
     private readonly TwitterClient _twitterClient;
-    private readonly OpenAIClient _openAiClient;
+    private readonly ITweetFilter _tweetFilter;
     private readonly string[] _accounts = new[] { "@example" };
 
-    public Worker(ILogger<Worker> logger)
+    public Worker(ILogger<Worker> logger, ITweetFilter tweetFilter)
     {
         _logger = logger;
+        _tweetFilter = tweetFilter;
         // In real code, keys should come from configuration or secrets
         _twitterClient = new TwitterClient(
             Environment.GetEnvironmentVariable("TWITTER_API_KEY"),
             Environment.GetEnvironmentVariable("TWITTER_API_SECRET"),
             Environment.GetEnvironmentVariable("TWITTER_ACCESS_TOKEN"),
             Environment.GetEnvironmentVariable("TWITTER_ACCESS_SECRET"));
-
-        var endpoint = Environment.GetEnvironmentVariable("OPENAI_ENDPOINT");
-        var key = Environment.GetEnvironmentVariable("OPENAI_KEY");
-        if (!string.IsNullOrEmpty(endpoint) && !string.IsNullOrEmpty(key))
-        {
-            _openAiClient = new OpenAIClient(new Uri(endpoint), new Azure.AzureKeyCredential(key));
-        }
     }
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -42,7 +35,7 @@ public class Worker : BackgroundService
                     var timeline = await _twitterClient.Timelines.GetUserTimelineAsync(user);
                     foreach (var tweet in timeline)
                     {
-                        if (await CheckTweetAsync(tweet))
+                        if (await _tweetFilter.ShouldProcessAsync(tweet))
                         {
                             await WriteResultAsync(tweet);
                         }
@@ -57,21 +50,6 @@ public class Worker : BackgroundService
         }
     }
 
-    private async Task<bool> CheckTweetAsync(ITweet tweet)
-    {
-        if (_openAiClient == null)
-        {
-            // Fallback simple keyword check
-            return tweet.FullText.Contains("AI", StringComparison.OrdinalIgnoreCase);
-        }
-
-        var prompt = $"Does this tweet mention AI or machine learning? {tweet.FullText}";
-        var response = await _openAiClient.GetCompletionsAsync(
-            Environment.GetEnvironmentVariable("OPENAI_DEPLOYMENT"),
-            prompt);
-        var text = response.Value.Choices[0].Text;
-        return text.Contains("yes", StringComparison.OrdinalIgnoreCase);
-    }
 
     private static async Task WriteResultAsync(ITweet tweet)
     {
